feat(desktop): move workdir selector into input, slim sidebar to new-chat button
- Workdir selector now lives at the top-left of the message input
  (DesktopWorkdirSelector, menu expands upward), shown only in the
  new-session state; it disappears once the conversation starts
- Sidebar reduced to title + 新对话 button; session history list
  removed for now (redesign planned)
- MessageInput gains a host-agnostic workdirSelector slot prop
- Update spec FR-005 / user story 2 and tests accordingly
