+++ docs/specs/ui/desktop-app.md
@@ -27,17 +27,18 @@
 
 ### 用户故事 2 - 会话管理（优先级：P2）
 
-作为用户，我希望在应用内通过侧边栏查看会话历史、切换会话或新建会话，以便我可以在不同任务之间切换。
+作为用户，我希望在应用内新建会话并恢复上次的会话与工作目录，以便我可以在不同任务之间切换。
 
 **为什么是这个优先级**：会话管理是日常使用的基础能力，且 sidebar 是 desktop 相对 IDE 插件的布局扩展。
 
-**独立测试**：进行两个会话后，通过侧边栏在它们之间切换，验证各自消息记录正确恢复。
+**独立测试**：进行一个会话后点击"新对话"开始全新会话；重启应用验证上次会话与工作目录恢复。
 
 **验收场景**：
 
-1. **假设**应用已有一个会话，**当**用户在侧边栏选择另一个历史会话，**则**应用必须在当前窗口内切换到该会话并恢复其消息记录。
-2. **假设**应用正在运行，**当**用户点击"新会话"，**则**必须在当前窗口内开始一个全新会话。
-3. **假设**用户关闭并重新打开应用，**当**应用启动，**则**必须恢复上次的会话与工作目录。
+1. **假设**应用正在运行，**当**用户点击侧边栏"新对话"，**则**必须在当前窗口内开始一个全新会话（回到欢迎页 + 输入框状态）。
+2. **假设**用户关闭并重新打开应用，**当**应用启动，**则**必须恢复上次的会话与工作目录。
+
+> 注：侧边栏会话历史列表暂时移除（后续有独立计划），历史会话能力由 CLI session store 保留。
 
 ---
 
@@ -153,7 +154,7 @@
 - **FR-002**：主进程必须通过 stdio 连接全局安装的 `wave-code` CLI（`wave --stdio`），而不是 fork 子进程运行 Agent。
 - **FR-003**：启动时必须检测 CLI 是否存在；缺失时必须通过 `npm install -g wave-code --registry=https://registry.npmmirror.com` 自动安装；Node.js 主版本低于 20 时必须提示升级 Node.js；Node.js/npm 完全缺失时必须提示安装 Node.js。
 - **FR-004**：启动时必须比较已安装 CLI 版本与要求的最低版本，低于最低版本时必须自动升级。
-- **FR-005**：UI 必须提供常驻侧边栏，包含工作目录选择入口与会话列表；侧边栏必须基于 webview 现有的 `SessionList` 抽取复用；会话切换与新建会话均在当前窗口内完成（单窗口架构）。工作目录选择入口为侧边栏顶部 workdir 区域的下拉选择器（与权限模式下拉同模式：自定义 trigger + 菜单、点外部关闭），菜单选项包含「最近工作目录」列表与「浏览…」项；无工作目录时侧边栏仍渲染，workdir 区域显示「选择工作目录…」占位并使用同一个下拉，首次启动与切换工作目录体验一致。
+- **FR-005**：UI 必须提供常驻侧边栏，侧边栏仅包含"新对话"按钮（plus 图标 + 文字），点击在当前窗口内开始全新会话（单窗口架构）；侧边栏暂不提供会话历史列表（后续有独立计划）。工作目录选择入口位于输入框（input-wrapper）内部左上角，为下拉选择器（与权限模式下拉同模式：自定义 trigger + 菜单、点外部关闭），菜单选项包含「最近工作目录」列表与「浏览…」项；该选择器**仅在新会话状态（消息列表为空）显示**，对话一旦开始即隐藏，工作目录在新会话期间确定；无工作目录时显示「选择工作目录…」占位，首次启动与切换工作目录体验一致。
 - **FR-006**：工作目录通过下拉里的「浏览…」项触发原生目录选择对话框选择；必须支持从最近目录列表中移除条目（仅删除持久化记录，不删除磁盘文件）；不再提供"仅临时聊天"临时目录选项。
 - **FR-007**：webview UI 必须零修改复用现有 Chat 组件；desktop 特有布局（侧边栏、工作目录选择）必须基于现有组件抽取扩展，通过 `Window.waveHostType` 区分宿主。
 - **FR-008**：确认（Confirmation）、AskUserQuestion、`@文件` 提及、图片粘贴、PDF 附件必须复用 webview 现有实现，desktop 仅实现对应的宿主消息处理（如 `listFiles` 请求），行为与 IDE 插件一致。
